requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
google-generativeai
pydantic
requests
beautifulsoup4
chromadb
python-dotenv
streamlit
tavily-python>=0.5.0
